chore: remove ts paths & add internal utils package
