--- .storybook/utils.ts
@@ -1,3 +1,0 @@
-export const renderStory = (story: any, ctx: any) => {
-  return story.render?.(story.args, ctx);
-};
