+++ apps/app/tsconfig.json
@@ -1,5 +1,10 @@
 {
   "extends": "../../tsconfig.json",
   "include": ["src"],
+  "compilerOptions": {
+    "paths": {
+      "~/*": ["./src/*"]
+    }
+  },
   "references": [{ "path": "./tsconfig.node.json" }]
 }
