+++ apps/default-app/package.json
@@ -37,7 +37,6 @@
     "typescript": "^5.4.2",
     "unocss": "^66.0.0",
     "vite": "^6.2.2",
-    "vite-plugin-css-injected-by-js": "^3.4.0",
-    "vite-tsconfig-paths": "^5.1.0"
+    "vite-plugin-css-injected-by-js": "^3.4.0"
   }
 }
