+++ package.json
@@ -121,7 +121,6 @@
     "usehooks-ts": "^2.9.1",
     "vite": "^6.2.2",
     "vite-plugin-dts": "^4.5.4",
-    "vite-tsconfig-paths": "^5.1.0",
     "vitest": "^3.0.9",
     "yup": "^1.3.2",
     "zod": "^3.22.4",
