+++ packages/core/src/Dialog/Dialog.stories.tsx
@@ -1,6 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import isChromatic from "chromatic/isChromatic";
 import { expect, userEvent, within } from "storybook/test";
+import { setupChromatic } from "@hitachivantara/internal";
 import {
   HvDialog,
   HvDialogActions,
@@ -13,7 +14,6 @@ import { FormStory } from "./stories/FormStory";
 import { LongContentStory } from "./stories/LongContentStory";
 import { MainStory } from "./stories/MainStory";
 import { SemanticVariantsStory } from "./stories/SemanticVariantsStory";
-import { setupChromatic } from ".storybook/setupChromatic";
 
 const meta: Meta<typeof HvDialog> = {
   title: "Components/Dialog",
