+++ packages/core/src/TimePicker/TimePicker.stories.tsx
@@ -1,13 +1,12 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import { expect, userEvent, within } from "storybook/test";
+import { setupChromatic } from "@hitachivantara/internal";
 import {
   HvButton,
   HvTimePicker,
   HvTimePickerProps,
   HvTimePickerValue,
 } from "@hitachivantara/uikit-react-core";
-
-import { setupChromatic } from ".storybook/setupChromatic";
 
 export default {
   title: "Components/Time Picker",
