+++ packages/core/src/stories/Icons.test.stories.tsx
@@ -1,12 +1,11 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
+import { setupChromatic } from "@hitachivantara/internal";
 import {
   CheckboxCheck,
   HvIconSprite,
   HvIconSpriteProps,
   icons,
 } from "@hitachivantara/uikit-react-icons";
-
-import { setupChromatic } from ".storybook/setupChromatic";
 
 export default {
   title: "Tests/Icons",
