+++ packages/viz/src/BaseChart/stories/utils.tsx
@@ -9,9 +9,3 @@ export const vizDecorator: Decorator = (Story) => (
     </HvPanel>
   </HvVizProvider>
 );
-
-export const confusionMatrixDecorator: Decorator = (Story) => (
-  <HvVizProvider>
-    <HvPanel className="flex flex-col">{Story()}</HvPanel>
-  </HvVizProvider>
-);
